requirements.txt
typer==0.12.5
google-genai
